Change file structure of bootstrap process.
This way one can setup the process (for example) to catch unhandled exceptions. Right now, this is only used to output StartUpException in a properly formatted way to the console.
